+++ app/models/connect.rb
@@ -1,4 +1,6 @@
 class Connect < ActiveRecord::Base
+  after_commit AuditLog
+
   belongs_to :user
   belongs_to :provider