+++ spec/factories/ios_locations.rb
@@ -1,0 +1,6 @@
+FactoryGirl.define do
+  factory :ios_location do
+    name 'berlin'
+    creator
+  end
+end